Sort out handleMessage functions
@LudwigBr this is my attempt to solve this issue.

Derived classes evaluate first, in the else clause the super class is called.

Each class also includes a conditional reschedule. If we ever need to schedule a new even in one of the subclasses, that's possible now.

I've also removed your response message. The problem is, that message objects didn't get disposed properly, we would need to delete them all the time - which is simply not needed because we can recycle the existing object in 90% of the time.