enable battery overdraw
When a copter is running out of battery, i.e. remaining reaches 0, it will continue it's cee and the overdrawn energy will be accumulated in a watch parameter. The remaining will never get below 0 though. Therefore when this particular node finally reaches a charging station, it'll start charging from 0 remaining energy.